requirements.txt
duckdb==1.1.3
chromadb==0.5.20
fastapi==0.115.6
langchain==0.3.8
langchain_core==0.3.25
langchain-chroma==0.1.4
langchain_openai==0.2.12
langchain_pinecone==0.2.0
langgraph==0.2.59
pandas==1.5.3
psycopg2-binary==2.9.10
pydantic==2.10.3
pymongo==4.10.1
python-dotenv==1.0.1
uvicorn==0.34.0
tabulate
langfuse==2.57.1